Extract mdx extension registry from utility types
The utility types can now be imported from `mdx/types`, whereas referencing `mdx` registers the file extensions.
